backend/requirements.txt
asgiref==3.6.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
firebase-admin==6.1.0
fonttools==4.39.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
jws==0.1.3
kiwisolver==1.4.4
matplotlib==3.7.1
msgpack==1.0.5
numpy==1.24.3
oauth2client==4.1.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase==3.0.27
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
